Remove reduce_range

+++ torchao/kernel/intmm.py
@@ -2,7 +2,7 @@ import itertools
 import os
 import torch
 
-from torchao.utils import TORCH_VERSION_AT_LEAST_2_2, TORCH_VERSION_AT_LEAST_2_4
+from torchao.utils import TORCH_VERSION_AT_LEAST_2_2
 
 try:
     # Only works for torch2.2 or newer.
@@ -56,9 +56,10 @@ if TORCH_VERSION_AT_LEAST_2_2:
         
         if device_cpu or bad_dimensions_for_cublas:
             # fallback path
-            return torch.matmul(input.cpu().to(torch.int32), mat2.cpu().to(torch.int32)).to(
+            # Compute in float instead of int32 because int32 matmul is not parallelized on CPU
+            return torch.matmul(input.cpu().to(torch.float), mat2.cpu().to(torch.float)).to(
                 input.device.type
-            )
+            ).to(torch.int32)
 
         # cublas paths
         if not mat2.is_contiguous():  # silently gives incorrect result without this
